delete distributed_run_multi_client.py and move impl in distributed_run.py
